// src/app/shared/models/detalleParametro.ts
export interface DetalleParametro{
        idDetaParametro?: number;
        idParametro?: number;
        deParametro?: string;
        coParametro?: string;
        deValor11?: string;
        deValor1?: string;
        deValor2?: string;
        deValor3?: string;
        deValor4?: string;
        feValor1?: string;
        feValor2?: string;
        feValor3?: string;
        feValor4?: string;
        indicadorBaja?: string;
        fechaCreacion?: string;
        usuarioModificacion?: string;
        ipModificacion?: string;
        fechaModificacion?: string;
        usuarioCreacion?: string;
        ipCreacion?: string;
}